Added the option for overriding the default watchdog observer.
This ability is useful when trying to watch over a mounted NFS,
in a linux machine, when the NFS is hosted on a windows machine, since
the default observer will use Inotify (watchdog assumes that it will be
the best polling method available on a linux machine) but Inotify syscall
won't work with the mounted NFS because it is hosted on windows.
In that case the user would want to override the default watchdog observer
with PollingObserver.

modified:   hachiko/hachiko.py

Signed-off-by: Eyal Royee <eyalroyee@gmail.com>